rename protocol_ctfCorrecion.py to protocol_ctfCorrection.py Adapt test to new mdoc usage replace comments with " by #
