+++ imod/protocols/__init__.py
@@ -34,7 +34,7 @@ from .protocol_tomoReconstruction import ProtImodTomoReconstruction
 from .protocol_applyTransformationMatrix import ProtImodApplyTransformationMatrix
 from .protocol_tsNormalization import ProtImodTSNormalization
 from .protocol_tomoNormalization import ProtImodTomoNormalization
-from .protocol_ctfCorrecion import ProtImodCtfCorrection
+from .protocol_ctfCorrection import ProtImodCtfCorrection
 from .protocol_ctfEstimation import ProtImodCtfEstimation
 from .protocol_tomoProjection import ProtImodTomoProjection
 from .protocol_goldBeadPicker3d import ProtImodGoldBeadPicker3d
