+++ imod/utils.py
@@ -251,27 +251,27 @@ def readCTFEstimationInfoFile(defocusFilePath, flag):
     tilt-series. These dictionaries are readable information for Scipion, useful to generate the corresponding output
     CTFTomoSeries object. """
 
-    " Read info as table "
+    # Read info as table
     ctfInfoIMODTable = readDefocusFileAsTable(defocusFilePath)
 
     if flag == 0:
-        " Plain estimation "
+        # Plain estimation
         return refactorCTFDefocusEstimationInfo(ctfInfoIMODTable)
 
     elif flag == 1:
-        " Astigmatism estimation "
+        # Astigmatism estimation
         return refactorCTFDesfocusAstigmatismEstimationInfo(ctfInfoIMODTable)
 
     elif flag == 4:
-        " Phase-shift estimation "
+        # Phase-shift estimation
         return refactorCTFDefocusPhaseShiftEstimationInfo(ctfInfoIMODTable)
 
     elif flag == 5:
-        " Astigmatism and phase shift estimation "
+        # Astigmatism and phase shift estimation
         return refactorCTFDefocusAstigmatismPhaseShiftEstimationInfo(ctfInfoIMODTable)
 
     elif flag == 37:
-        " Astigmatism, phase shift and cut-on frequency estimation "
+        # Astigmatism, phase shift and cut-on frequency estimation
         return refactorCTFDefocusAstigmatismPhaseShiftCutOnFreqEstimationInfo(ctfInfoIMODTable)
 
     else:
@@ -289,7 +289,7 @@ def refactorCTFDefocusEstimationInfo(ctfInfoIMODTable):
 
         for element in ctfInfoIMODTable:
 
-            " Segregate information from range"
+            # Segregate information from range
             for index in range(int(element[0]), int(element[1]) + 1):
                 if index in defocusUDict.keys():
                     defocusUDict[index].append(pwobj.Float(element[4]))
@@ -314,22 +314,22 @@ def refactorCTFDesfocusAstigmatismEstimationInfo(ctfInfoIMODTable):
 
         for element in ctfInfoIMODTable:
 
-            " Segregate information from range"
+            # Segregate information from range
             for index in range(int(element[0]), int(element[1]) + 1):
 
-                " Defocus U info "
+                # Defocus U info
                 if index in defocusUDict.keys():
                     defocusUDict[index].append(pwobj.Float(element[4]))
                 else:
                     defocusUDict[index] = [pwobj.Float(element[4])]
 
-                " Defocus V info "
+                # Defocus V info
                 if index in defocusVDict.keys():
                     defocusVDict[index].append(pwobj.Float(element[5]))
                 else:
                     defocusVDict[index] = [pwobj.Float(element[5])]
 
-                " Defocus angle info "
+                # Defocus angle info
                 if index in defocusAngleDict.keys():
                     defocusAngleDict[index].append(pwobj.Float(element[6]))
                 else:
@@ -352,16 +352,16 @@ def refactorCTFDefocusPhaseShiftEstimationInfo(ctfInfoIMODTable):
 
         for element in ctfInfoIMODTable:
 
-            " Segregate information from range"
+            # Segregate information from range
             for index in range(int(element[0]), int(element[1]) + 1):
 
-                " Defocus U info "
+                # Defocus U info
                 if index in defocusUDict.keys():
                     defocusUDict[index].append(pwobj.Float(element[4]))
                 else:
                     defocusUDict[index] = [pwobj.Float(element[4])]
 
-                " Phase shift info "
+                # Phase shift info
                 if index in phaseShiftDict.keys():
                     phaseShiftDict[index].append(pwobj.Float(element[5]))
                 else:
@@ -387,28 +387,28 @@ def refactorCTFDefocusAstigmatismPhaseShiftEstimationInfo(ctfInfoIMODTable):
 
         for element in ctfInfoIMODTable:
 
-            " Segregate information from range"
+            # Segregate information from range
             for index in range(int(element[0]), int(element[1]) + 1):
 
-                " Defocus U info "
+                # Defocus U info
                 if index in defocusUDict.keys():
                     defocusUDict[index].append(pwobj.Float(element[4]))
                 else:
                     defocusUDict[index] = [pwobj.Float(element[4])]
 
-                " Defocus V info "
+                # Defocus V info
                 if index in defocusVDict.keys():
                     defocusVDict[index].append(pwobj.Float(element[5]))
                 else:
                     defocusVDict[index] = [pwobj.Float(element[5])]
 
-                " Defocus angle info "
+                # Defocus angle info
                 if index in defocusAngleDict.keys():
                     defocusAngleDict[index].append(pwobj.Float(element[6]))
                 else:
                     defocusAngleDict[index] = [pwobj.Float(element[6])]
 
-                " Phase shift info "
+                # Phase shift info
                 if index in phaseShiftDict.keys():
                     phaseShiftDict[index].append(pwobj.Float(element[7]))
                 else:
@@ -435,34 +435,34 @@ def refactorCTFDefocusAstigmatismPhaseShiftCutOnFreqEstimationInfo(ctfInfoIMODTa
 
         for element in ctfInfoIMODTable:
 
-            " Segregate information from range"
+            # Segregate information from range
             for index in range(int(element[0]), int(element[1]) + 1):
 
-                " Defocus U info "
+                # Defocus U info
                 if index in defocusUDict.keys():
                     defocusUDict[index].append(pwobj.Float(element[4]))
                 else:
                     defocusUDict[index] = [pwobj.Float(element[4])]
 
-                " Defocus V info "
+                # Defocus V info
                 if index in defocusVDict.keys():
                     defocusVDict[index].append(pwobj.Float(element[5]))
                 else:
                     defocusVDict[index] = [pwobj.Float(element[5])]
 
-                " Defocus angle info "
+                # Defocus angle info
                 if index in defocusAngleDict.keys():
                     defocusAngleDict[index].append(pwobj.Float(element[6]))
                 else:
                     defocusAngleDict[index] = [pwobj.Float(element[6])]
 
-                " Phase shift info "
+                # Phase shift info
                 if index in phaseShiftDict.keys():
                     phaseShiftDict[index].append(pwobj.Float(element[7]))
                 else:
                     phaseShiftDict[index] = [pwobj.Float(element[7])]
 
-                " Cut-on frequency info "
+                # Cut-on frequency info
                 if index in cutOnFreqDict.keys():
                     cutOnFreqDict[index].append(pwobj.Float(element[8]))
                 else:
@@ -481,16 +481,16 @@ def generateDefocusIMODFileFromObject(ctfTomoSeries, defocusFilePath):
 
     tiltSeries = ctfTomoSeries.getTiltSeries()
 
-    " Check if there is CTF estimation information as list "
+    # Check if there is CTF estimation information as list
     if ctfTomoSeries.getFirstItem().hasEstimationInfoAsList():
 
         flag = ctfTomoSeries.getIMODDefocusFileFlag()
 
         if flag == 0:
-            " Plain estimation "
+            # Plain estimation
             defocusUDict = generateDefocusUDictionary(ctfTomoSeries)
 
-            " Write IMOD defocus file "
+            # Write IMOD defocus file
             with open(defocusFilePath, 'w') as f:
                 lines = []
 
@@ -499,31 +499,31 @@ def generateDefocusIMODFileFromObject(ctfTomoSeries, defocusFilePath):
                     if index + ctfTomoSeries.getNumberOfEstimationsInRange() > len(defocusUDict.keys()):
                         break
 
-                    " Dictionary keys is reversed because IMOD set indexes upside down Scipion (highest index for "
-                    " the tilt-image with the highest negative angle) "
+                    itemId = index-1
+
                     newLine = ("%d\t%d\t%.2f\t%.2f\t%d\n" % (
                         index,
                         index + ctfTomoSeries.getNumberOfEstimationsInRange(),
-                        round(tiltSeries[index + ctfTomoSeries.getNumberOfEstimationsInRange()].getTiltAngle(), 2),
-                        round(tiltSeries[index].getTiltAngle(), 2),
+                        round(tiltSeries[itemId + ctfTomoSeries.getNumberOfEstimationsInRange()].getTiltAngle(), 2),
+                        round(tiltSeries[itemId].getTiltAngle(), 2),
                         int(float(defocusUDict[index][0]))
                     ))
 
                     lines = [newLine] + lines
 
-                "Finally, add flag to the first line of file"
+                # Finally, add flag to the first line of file
                 lines[0] = lines[0][0:-1] + "\t2\n"
 
                 f.writelines(lines)
 
         elif flag == 1:
-            " Astigmatism estimation "
+            # Astigmatism estimation
 
             defocusUDict = generateDefocusUDictionary(ctfTomoSeries)
             defocusVDict = generateDefocusVDictionary(ctfTomoSeries)
             defocusAngleDict = generateDefocusAngleDictionary(ctfTomoSeries)
 
-            " Write IMOD defocus file "
+            # Write IMOD defocus file
             with open(defocusFilePath, 'w') as f:
                 lines = []
 
@@ -532,8 +532,8 @@ def generateDefocusIMODFileFromObject(ctfTomoSeries, defocusFilePath):
                     if index + ctfTomoSeries.getNumberOfEstimationsInRange() > len(defocusUDict.keys()):
                         break
 
-                    " Dictionary keys is reversed because IMOD set indexes upside down Scipion (highest index for "
-                    " the tilt-image with the highest negative angle) "
+                    # Dictionary keys is reversed because IMOD set indexes upside down Scipion (highest index for
+                    # the tilt-image with the highest negative angle)
                     newLine = ("%d\t%d\t%.2f\t%.2f\t%.1f\t%.1f\t%.2f\n" % (
                         index,
                         index + ctfTomoSeries.getNumberOfEstimationsInRange(),
@@ -546,18 +546,18 @@ def generateDefocusIMODFileFromObject(ctfTomoSeries, defocusFilePath):
 
                     lines = [newLine] + lines
 
-                " This line is added at the beginning of the file in order to match the IMOD defocus file format. "
+                # This line is added at the beginning of the file in order to match the IMOD defocus file format.
                 lines = ["1\t0\t0.0\t0.0\t0.0\t3\n"] + lines
 
                 f.writelines(lines)
 
         elif flag == 4:
-            " Phase-shift estimation "
+            # Phase-shift estimation
 
             defocusUDict = generateDefocusUDictionary(ctfTomoSeries)
             phaseShiftDict = generatePhaseShiftDictionary(ctfTomoSeries)
 
-            " Write IMOD defocus file "
+            # Write IMOD defocus file
             with open(defocusFilePath, 'w') as f:
                 lines = []
 
@@ -566,8 +566,8 @@ def generateDefocusIMODFileFromObject(ctfTomoSeries, defocusFilePath):
                     if index + ctfTomoSeries.getNumberOfEstimationsInRange() > len(defocusUDict.keys()):
                         break
 
-                    " Dictionary keys is reversed because IMOD set indexes upside down Scipion (highest index for "
-                    " the tilt-image with the highest negative angle) "
+                    # Dictionary keys is reversed because IMOD set indexes upside down Scipion (highest index for
+                    # the tilt-image with the highest negative angle)
                     newLine = ("%d\t%d\t%.2f\t%.2f\t%.1f\t%.2f\n" % (
                         index,
                         index + ctfTomoSeries.getNumberOfEstimationsInRange(),
@@ -579,20 +579,20 @@ def generateDefocusIMODFileFromObject(ctfTomoSeries, defocusFilePath):
 
                     lines = [newLine] + lines
 
-                " This line is added at the beginning of the file in order to match the IMOD defocus file format. "
+                # This line is added at the beginning of the file in order to match the IMOD defocus file format.
                 lines = ["4\t0\t0.0\t0.0\t0.0\t3\n"] + lines
 
                 f.writelines(lines)
 
         elif flag == 5:
-            " Astigmatism and phase shift estimation "
+            # Astigmatism and phase shift estimation
 
             defocusUDict = generateDefocusUDictionary(ctfTomoSeries)
             defocusVDict = generateDefocusVDictionary(ctfTomoSeries)
             defocusAngleDict = generateDefocusAngleDictionary(ctfTomoSeries)
             phaseShiftDict = generatePhaseShiftDictionary(ctfTomoSeries)
 
-            " Write IMOD defocus file "
+            # Write IMOD defocus file
             with open(defocusFilePath, 'w') as f:
                 lines = []
 
@@ -601,8 +601,8 @@ def generateDefocusIMODFileFromObject(ctfTomoSeries, defocusFilePath):
                     if index + ctfTomoSeries.getNumberOfEstimationsInRange() > len(defocusUDict.keys()):
                         break
 
-                    " Dictionary keys is reversed because IMOD set indexes upside down Scipion (highest index for "
-                    " the tilt-image with the highest negative angle) "
+                    # Dictionary keys is reversed because IMOD set indexes upside down Scipion (highest index for
+                    # the tilt-image with the highest negative angle)
                     newLine = ("%d\t%d\t%.2f\t%.2f\t%.1f\t%.1f\t%.2f\t%.2f\n" % (
                         index,
                         index + ctfTomoSeries.getNumberOfEstimationsInRange(),
@@ -616,13 +616,13 @@ def generateDefocusIMODFileFromObject(ctfTomoSeries, defocusFilePath):
 
                     lines = [newLine] + lines
 
-                " This line is added at the beginning of the file in order to match the IMOD defocus file format "
+                # This line is added at the beginning of the file in order to match the IMOD defocus file format
                 lines = ["5\t0\t0.0\t0.0\t0.0\t3\n"] + lines
 
                 f.writelines(lines)
 
         elif flag == 37:
-            " Astigmatism, phase shift and cut-on frequency estimation "
+            # Astigmatism, phase shift and cut-on frequency estimation
 
             defocusUDict = generateDefocusUDictionary(ctfTomoSeries)
             defocusVDict = generateDefocusVDictionary(ctfTomoSeries)
@@ -630,7 +630,7 @@ def generateDefocusIMODFileFromObject(ctfTomoSeries, defocusFilePath):
             phaseShiftDict = generatePhaseShiftDictionary(ctfTomoSeries)
             cutOnFreqDict = generateCutOnFreqDictionary(ctfTomoSeries)
 
-            " Write IMOD defocus file "
+            # Write IMOD defocus file
             with open(defocusFilePath, 'w') as f:
                 lines = []
 
@@ -639,8 +639,8 @@ def generateDefocusIMODFileFromObject(ctfTomoSeries, defocusFilePath):
                     if index + ctfTomoSeries.getNumberOfEstimationsInRange() > len(defocusUDict.keys()):
                         break
 
-                    " Dictionary keys is reversed because IMOD set indexes upside down Scipion (highest index for "
-                    " the tilt-image with the highest negative angle) "
+                    # Dictionary keys is reversed because IMOD set indexes upside down Scipion (highest index for
+                    # the tilt-image with the highest negative angle)
                     newLine = ("%d\t%d\t%.2f\t%.2f\t%.1f\t%.1f\t%.2f\t%.2f\t%.4f\n" % (
                         index,
                         index + ctfTomoSeries.getNumberOfEstimationsInRange(),
@@ -655,7 +655,7 @@ def generateDefocusIMODFileFromObject(ctfTomoSeries, defocusFilePath):
 
                     lines = [newLine] + lines
 
-                " This line is added at the beginning of the file in order to match the IMOD defocus file format "
+                # This line is added at the beginning of the file in order to match the IMOD defocus file format
                 lines = ["37\t0\t0.0\t0.0\t0.0\t3\n"] + lines
 
                 f.writelines(lines)
@@ -665,13 +665,13 @@ def generateDefocusIMODFileFromObject(ctfTomoSeries, defocusFilePath):
                             "1, 4, 5, and 37.")
 
     else:
-        " There is no information available as list (not an IMOD CTF estimation) "
+        # There is no information available as list (not an IMOD CTF estimation)
 
         with open(defocusFilePath, 'w') as f:
             lines = []
 
-            " CtfTomoSeries is iterated inversely because IMOD set indexes upside down Scipion (highest index for "
-            " the tilt-image with the highest negative angle) "
+            # CtfTomoSeries is iterated inversely because IMOD set indexes upside down Scipion (highest index for
+            # the tilt-image with the highest negative angle)
             for ctfTomo in ctfTomoSeries:
                 index = ctfTomo.getIndex()
